Default comparator extension tests
These verify that extension works as intended. The tests are a bit
implementation-specific at the point when two extension types are
compared: the SRFI does not specify any explicit order between the
registered types (and the method of registering itself). So I have
defined it to be the order of registering.

The TODO got deleted because these tests should go into SRFI 113
itself to avoid creating a circular dependency between libraries.