Added extension registry for the default comparator
Other system libraries can call register-default-comparator! to add
their type-specific comparator to the global registry. The intended
usage of this procedure is limited to the phase when libraries are
loaded. As Scheme does not have a standard way to unload a library,
no means are provided to unregister a comparator.