Exported register-default-comparator!
The libraries got shuffled a bit to cleanly define the public interface
for the users, the private interface for the system, and the actual library
as a collection of procedures and global data.

(srfi 114) is the user-visible interface, as defined by the SRFI 114.
Anything below it is an implementation-private interface, namely:

  - (srfi 114 export) contains all symbols exported by the library

  - (srfi 114 default-comparator) exports register-default-comparator!

So, the user-level code would import just (srfi 114), while an internal
library that wishes to register some of its types within the default
comparator would import (srfi 114) to define the comparators and
(srfi 114 default-comparator) to register them.