Proper hash functions from compound types
car/cdr, pairs, and improper lists are obvious.

List and vector hash functions actually do not depend on the order
of elements. It was not explicitly intended, it is just easier to
do so, and this preserves all required hash function properties.

Lists and vectors have a limit on the number of elements visited.
This is done to keep hash function calculation time roughly O(1).

Also they have 'weird' starting values. This is done to prevent
contention around the '0' hash value for all empty/zero/null/etc.

There are no tests because I am too lazy to think up some special
element hash functions to properly test differences between them
and the default hash which is necessary to confirm that comparator
constructors actually use custom hashes instead of the default one.